Fixing home screen issues
- changes related to comments in PR
